new scaffold

+++ OnlineStore/Controllers/BrandsController.cs
@@ -1,5 +1,9 @@
-﻿#nullable disable
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using OnlineStore.Models;
 
@@ -17,13 +21,15 @@ namespace OnlineStore.Controllers
         // GET: Brands
         public async Task<IActionResult> Index()
         {
-            return View(await _context.Brands.ToListAsync());
+              return _context.Brands != null ? 
+                          View(await _context.Brands.ToListAsync()) :
+                          Problem("Entity set 'OnlineStoreContext.Brands'  is null.");
         }
 
         // GET: Brands/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
+            if (id == null || _context.Brands == null)
             {
                 return NotFound();
             }
@@ -63,7 +69,7 @@ namespace OnlineStore.Controllers
         // GET: Brands/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
+            if (id == null || _context.Brands == null)
             {
                 return NotFound();
             }
@@ -114,7 +120,7 @@ namespace OnlineStore.Controllers
         // GET: Brands/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
+            if (id == null || _context.Brands == null)
             {
                 return NotFound();
             }
@@ -134,15 +140,23 @@ namespace OnlineStore.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (_context.Brands == null)
+            {
+                return Problem("Entity set 'OnlineStoreContext.Brands'  is null.");
+            }
             var brand = await _context.Brands.FindAsync(id);
-            _context.Brands.Remove(brand);
+            if (brand != null)
+            {
+                _context.Brands.Remove(brand);
+            }
+            
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
 
         private bool BrandExists(int id)
         {
-            return _context.Brands.Any(e => e.BrandId == id);
+          return (_context.Brands?.Any(e => e.BrandId == id)).GetValueOrDefault();
         }
     }
 }
