+++ OnlineStore/Controllers/StocksController.cs
@@ -1,4 +1,7 @@
-﻿#nullable disable
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
@@ -25,7 +28,7 @@ namespace OnlineStore.Controllers
         // GET: Stocks/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
+            if (id == null || _context.Stocks == null)
             {
                 return NotFound();
             }
@@ -71,7 +74,7 @@ namespace OnlineStore.Controllers
         // GET: Stocks/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
+            if (id == null || _context.Stocks == null)
             {
                 return NotFound();
             }
@@ -126,7 +129,7 @@ namespace OnlineStore.Controllers
         // GET: Stocks/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
+            if (id == null || _context.Stocks == null)
             {
                 return NotFound();
             }
@@ -148,15 +151,23 @@ namespace OnlineStore.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (_context.Stocks == null)
+            {
+                return Problem("Entity set 'OnlineStoreContext.Stocks'  is null.");
+            }
             var stock = await _context.Stocks.FindAsync(id);
-            _context.Stocks.Remove(stock);
+            if (stock != null)
+            {
+                _context.Stocks.Remove(stock);
+            }
+            
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
 
         private bool StockExists(int id)
         {
-            return _context.Stocks.Any(e => e.StoreId == id);
+          return (_context.Stocks?.Any(e => e.StoreId == id)).GetValueOrDefault();
         }
     }
 }
