+++ OnlineStore/Controllers/StoresController.cs
@@ -1,5 +1,9 @@
-﻿#nullable disable
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using OnlineStore.Models;
 
@@ -17,13 +21,15 @@ namespace OnlineStore.Controllers
         // GET: Stores
         public async Task<IActionResult> Index()
         {
-            return View(await _context.Stores.ToListAsync());
+              return _context.Stores != null ? 
+                          View(await _context.Stores.ToListAsync()) :
+                          Problem("Entity set 'OnlineStoreContext.Stores'  is null.");
         }
 
         // GET: Stores/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
+            if (id == null || _context.Stores == null)
             {
                 return NotFound();
             }
@@ -63,7 +69,7 @@ namespace OnlineStore.Controllers
         // GET: Stores/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
+            if (id == null || _context.Stores == null)
             {
                 return NotFound();
             }
@@ -114,7 +120,7 @@ namespace OnlineStore.Controllers
         // GET: Stores/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
+            if (id == null || _context.Stores == null)
             {
                 return NotFound();
             }
@@ -134,15 +140,23 @@ namespace OnlineStore.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (_context.Stores == null)
+            {
+                return Problem("Entity set 'OnlineStoreContext.Stores'  is null.");
+            }
             var store = await _context.Stores.FindAsync(id);
-            _context.Stores.Remove(store);
+            if (store != null)
+            {
+                _context.Stores.Remove(store);
+            }
+            
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
 
         private bool StoreExists(int id)
         {
-            return _context.Stores.Any(e => e.StoreId == id);
+          return (_context.Stores?.Any(e => e.StoreId == id)).GetValueOrDefault();
         }
     }
 }
